backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
arabic-reshaper==3.0.0
asn1crypto==1.5.1
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
cssselect2==0.8.0
distro==1.9.0
dotenv==0.9.9
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.61.1
freetype-py==2.5.1
google-auth==2.47.0
google-genai==1.57.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
oscrypto==1.3.0
packaging==25.0
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.29.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydyf==0.12.1
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
pypdf==6.6.0
pypdfium2==5.3.0
pyphen==0.17.2
pytesseract==0.3.13
python-bidi==0.6.7
python-docx==1.2.0
python-dotenv==1.2.1
python-pptx==1.0.2
PyYAML==6.0.3
regex==2025.11.3
reportlab==4.4.9
requests==2.32.5
rlPyCairo==0.4.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
svglib==1.6.0
tenacity==9.1.2
tiktoken==0.12.0
tinycss2==1.5.1
tinyhtml5==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritools==6.0.1
urllib3==2.6.3
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.5
xhtml2pdf==0.2.17
xlsxwriter==3.2.9
zopfli==0.4.0
